Move transformFromDate into the TimeZoneService
